Section 26 - Authorization, Policies, Gates

+++ app/Http/Controllers/PostsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
+use Illuminate\Support\Facades\Gate;
 // use Illuminate\Support\Facades\DB;
 
 class PostsController extends Controller
@@ -83,7 +84,15 @@ class PostsController extends Controller
    */
   public function edit($id)
   {
-    return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+    $post = BlogPost::findOrFail($id);
+
+    // if (Gate::denies('posts.update', $post)) {
+    //   abort(403, "You can't edit this blog post");
+    // }
+
+    $this->authorize($post);
+
+    return view('posts.edit', ['post' => $post]);
   }
 
   /**
@@ -96,6 +105,9 @@ class PostsController extends Controller
   public function update(StorePost $request, $id)
   {
     $post = BlogPost::findOrFail($id);
+
+    $this->authorize($post);
+
     $validated = $request->validated();
 
     $post->fill($validated);
@@ -114,6 +126,13 @@ class PostsController extends Controller
   public function destroy($id)
   {
     $post = BlogPost::findOrFail($id);
+
+    // if (Gate::denies('posts.delete', $post)) {
+    //   abort(403, "You can't delete this blog post");
+    // }
+
+    $this->authorize($post);
+
     $post->delete();
 
     session()->flash('status', 'The blog post was deleted!');
